check_http: Move test cases
CMK-14257
